requirements.txt
PyGithub>=1.59.0
requests>=2.28.0
PyYAML>=6.0
